pyproject.toml
[project]
name = "voiceagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents==0.8.7",
    "livekit-plugins-openai==0.8.1",
    "livekit-plugins-silero==0.6.4",
    "python-dotenv~=1.0",
]
